added GET with all parameters and with filters using webFlux

+++ src/main/kotlin/ghar/backend/kotlin/spirng5docker/model/Employee.kt
@@ -4,6 +4,6 @@ class Employee(val id: Int,
                val name: String,
                val age: Int,
                val department: String,
-               val salary: Int) {
+               val salary: Double) {
 
 }
